Backend/requirements.txt
asgiref==3.10.0
black==25.9.0
click==8.3.0
colorama==0.4.6
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
PyJWT==2.10.1
python-decouple==3.8
pytokens==0.1.10
sqlparse==0.5.3
tzdata==2025.2
